move constants to primitives-traits
